Allow single process access to multiple repos.
These changes modify the behavior of the Environment and hdf5
FileHandler singleton classes to instantiate singleton instances based
on the path of the checked out repository. This seems to fix all
outstanding bugs related to these odd behaviors, but is untested as of
yet.

The TransactionRegister singleton did not need to be updated because it
the reference counting scheme is isolated for each lmdb environment
which is passed to it. As different repository instances in the same
process will now have unique lmdb environment handles, it does not
require any modifications to it's current behavior.