Added CLI checkout and commit option. Fixed checkout bug.
Checkout option is fully tested. Commit option still needs tests.

Checkout bug fixed. In cases where you checked out a branch on the CLI
for writing, and then opened a writer checkout in the API (without
specifying a branch), the default branch arg in the repo.checkout method
would reset the branch to `master`. Now, if no branch is provided, we
will use the branch of the latest writer checkout staging area. Tests
were updated to reflect this new behavior